Use sassc compatible gems
The middleman update in 14a9a20781e33d732c2d26fbcb10105baa5d066d
also ended up introducing sassc as a dependency. The versions of bourbon
and neat previously in the Gemfile do not work with sassc,
and https://baccano.io/ currently shows a SassC::SyntaxError on the
page. The latest versions of bourbon and neat should work fine with
sassc.